Undo binary breaking change
