--- src/OrchardCore/OrchardCore.DisplayManagement/ServiceCollectionExtensions.cs
@@ -1,15 +1,0 @@
-using Microsoft.Extensions.DependencyInjection;
-using OrchardCore.Settings;
-
-namespace OrchardCore.DisplayManagement.Handlers;
-
-public static class ServiceCollectionExtensions
-{
-    public static IServiceCollection AddSiteDisplayDriver<TDriver>(this IServiceCollection services)
-        where TDriver : class, IDisplayDriver<ISite>
-    => services.AddDisplayDriver<ISite, TDriver>();
-
-    public static IServiceCollection AddDisplayDriver<TModel, TDriver>(this IServiceCollection services)
-        where TDriver : class, IDisplayDriver<TModel>
-    => services.AddScoped<IDisplayDriver<TModel>, TDriver>();
-}
